requirements.txt

asgiref==3.8.1
bcrypt==3.2.0
beautifulsoup4==4.10.0
blinker==1.4
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
click==8.0.3
colorama==0.4.4
contourpy==1.2.1
cowsay==6.1
cryptography==3.4.8
cycler==0.12.1
distlib==0.3.8
distro==1.7.0
Django==5.0.7
djangorestframework==3.15.2
exceptiongroup==1.1.3
fasteners==0.14.1
filelock==3.15.4
fonttools==4.53.1
future==0.18.2
html5lib==1.1
httplib2==0.20.2
importlib-metadata==4.6.4
iniconfig==2.0.0
jeepney==0.7.1
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
meson==0.61.2
monotonic==1.6
more-itertools==8.10.0
numpy==2.0.0
oauthlib==3.2.0
olefile==0.46
packaging==23.2
pandas==2.2.2
paramiko==2.9.3
pbr==6.0.0
pexpect==4.8.0
platformdirs==4.2.2
pluggy==1.3.0
protobuf==3.12.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycparser==2.22
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==7.4.2
pytz==2022.1
pyxdg==0.27
requests==2.25.1
rest_framework_simplejwt==0.0.2
scour==0.38.2
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.5.1
ssh-import-id==5.11
urllib3==1.26.5
virtualenv==20.26.3
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.0
wadllib==1.3.6
webencodings==0.5.1
zipp==1.0.0
